ai_conversation_engine/requirements.txt
blinker
certifi
charset-normalizer
click
greenlet
idna
itsdangerous
Jinja2
MarkupSafe
typing_extensions
urllib3
python-dotenv
beautifulsoup4
openai
rapidfuzz
httpx
tenacity
google-generativeai
redis
Quart
quart-cors
hypercorn
pytest
pytest-asyncio
prometheus-client
pydantic
structlog
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
hypothesis
locust